Stop Mage_Dataflow from uploading CSVs to S3
Mage_Dataflow is attempting to sync uploaded CSV files from the var folder to
S3. This is causing an exception as Magento assumes that you only ever want to
upload stuff from the media folder.

This change stops Mage_Dataflow from syncing CSVs to S3. Everything is saved to
the local filesystem instead.